refactor: better core naming
