+++ cpast/core.py
